+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,8 +1,10 @@
 package janggi.domain.piece;
 
-import janggi.domain.RawRoute;
-import janggi.domain.Route;
+import janggi.domain.piece.direction.RawRoute;
+import janggi.domain.piece.direction.Route;
 import janggi.domain.Team;
+import janggi.domain.piece.direction.Position;
+import janggi.domain.piece.direction.RawPosition;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
