+++ src/main/java/janggi/domain/piece/Soldier.java
@@ -3,8 +3,10 @@ package janggi.domain.piece;
 import static janggi.domain.Team.BLUE;
 import static janggi.domain.Team.RED;
 
-import janggi.domain.RawRoute;
+import janggi.domain.piece.direction.RawRoute;
 import janggi.domain.Team;
+import janggi.domain.piece.direction.Position;
+import janggi.domain.piece.direction.RawPosition;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
