+++ src/main/java/janggi/domain/piece/direction/RawPosition.java
@@ -1,4 +1,4 @@
-package janggi.domain.piece;
+package janggi.domain.piece.direction;
 
 public record RawPosition(int x, int y) {
