+++ src/main/java/janggi/domain/piece/direction/RawRoute.java
@@ -1,6 +1,5 @@
-package janggi.domain;
+package janggi.domain.piece.direction;
 
-import janggi.domain.piece.RawPosition;
 import java.util.List;
 
 public record RawRoute(List<RawPosition> rawPositions) {
