+++ src/test/java/janggi/domain/PiecesTest.java
@@ -9,8 +9,9 @@ import janggi.domain.piece.Cannon;
 import janggi.domain.piece.Chariot;
 import janggi.domain.piece.Horse;
 import janggi.domain.piece.Piece;
-import janggi.domain.piece.Position;
+import janggi.domain.piece.direction.Position;
 import janggi.domain.piece.Soldier;
+import janggi.domain.piece.direction.Route;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
