+++ src/test/java/janggi/domain/piece/ElephantTest.java
@@ -3,7 +3,8 @@ package janggi.domain.piece;
 import static janggi.domain.Team.RED;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.domain.Route;
+import janggi.domain.piece.direction.Route;
+import janggi.domain.piece.direction.Position;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
